bit manipulation

+++ 0x14-bit_manipulation/2-get_bit.c
@@ -6,7 +6,7 @@
  * @n: unsigned long integer to get bit
  * Return: return value of bit at index else -1 if error
  */
-int get_bit(unsigned int n, unsigned int index)
+int get_bit(unsigned long int n, unsigned int index)
 {
 	if (index >= (sizeof(unsigned long int) * 8))
 		return (0);
